requirements.txt
numpy
matplotlib
mmpycocotools
numpy
six
terminaltables
timm
opencv-python
imageio
scipy
pycocotools
Pillow
scikit-learn
text2vec
setuptools
gradio==3.41.2
openai